requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.11
qdrant-client>=1.6.0
pypdf>=3.17.0
openai>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.1
streamlit>=1.28.0
